add userLogin

+++ app/Http/Controllers/TasksController.php
@@ -14,13 +14,23 @@ class TasksController extends Controller
      */
     public function index()
     {
-        // タスク一覧を取得
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
         
-        // タスク一覧ビューで表示
-        return view('tasks.index',[
-            'tasks' => $tasks,
-        ]);
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // タスク一覧を取得
+            //$tasks = Task::all();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+        
+            $data = [
+                    'user' => $user,
+                    'tasks' => $tasks,
+            ];
+        }
+        
+        // Welcomeビューで表示
+        return view('welcome', $data);
     }
 
     /**
@@ -51,11 +61,11 @@ class TasksController extends Controller
             'content' => 'required',
             'status' => 'required|max:10',
         ]);
-        // メッセージを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -69,13 +79,21 @@ class TasksController extends Controller
      */
     public function show($id)
     {
+
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-
-        // メッセージ詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+            
+        // 本人
+        if (\Auth::id() === $task->user_id) {
+        
+            // メッセージ詳細ビューでそれを表示
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }
+        
+        return redirect('/');
+        
     }
 
     /**
@@ -89,10 +107,17 @@ class TasksController extends Controller
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
 
-        // メッセージ編集ビューでそれを表示
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        // 本人
+        if (\Auth::id() === $task->user_id) {
+        
+            // メッセージ編集ビューでそれを表示
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);            
+        }
+        
+         return redirect('/');
+
     }
 
     /**
@@ -128,10 +153,13 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
+
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
